# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hangout Alley Siargao Restaurant Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tables:
    get:
      operationId: getTables
      tags: [tables]
      summary: List all tables
      responses:
        "200":
          description: List of tables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Table"
    post:
      operationId: createTable
      tags: [tables]
      summary: Create a table
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTableBody"
      responses:
        "201":
          description: Table created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
  /tables/{id}:
    put:
      operationId: updateTable
      tags: [tables]
      summary: Update a table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTableBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"

  # ORDERS
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: date
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # RESERVATIONS
  /reservations:
    get:
      operationId: getReservations
      tags: [reservations]
      summary: List all reservations
      parameters:
        - name: date
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Create a reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationBody"
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /reservations/{id}:
    get:
      operationId: getReservation
      tags: [reservations]
      summary: Get a reservation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /reservations/{id}/status:
    patch:
      operationId: updateReservationStatus
      tags: [reservations]
      summary: Update reservation status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReservationStatusBody"
      responses:
        "200":
          description: Updated reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"

  # REVIEWS
...
  /staff:
    get:
      operationId: getStaff
      tags: [staff]
      summary: List all staff members
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: createStaffMember
      tags: [staff]
      summary: Add a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffMemberBody"
      responses:
        "201":
          description: Staff member created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{id}:
    put:
      operationId: updateStaffMember
      tags: [staff]
      summary: Update a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffMemberBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
    delete:
      operationId: deleteStaffMember
      tags: [staff]
      summary: Remove a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ANALYTICS
...
components:
  schemas:
...
    Table:
      type: object
      properties:
        id:
          type: integer
        tableNumber:
          type: string
        capacity:
          type: integer
        location:
          type: ["string", "null"]
        isActive:
          type: boolean
      required: [id, tableNumber, capacity, isActive]
    CreateTableBody:
      type: object
      properties:
        tableNumber:
          type: string
        capacity:
          type: integer
        location:
          type: ["string", "null"]
        isActive:
          type: boolean
      required: [tableNumber, capacity, isActive]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        menuItemId:
          type: integer
        menuItemName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
        notes:
          type: ["string", "null"]
      required: [id, menuItemId, menuItemName, quantity, unitPrice, subtotal]
    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: ["string", "null"]
        type:
          type: string
          enum: [dine_in, takeout]
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, completed, cancelled]
        tableId:
          type: ["integer", "null"]
        tableNumber:
          type: ["string", "null"]
        totalAmount:
          type: number
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, customerName, type, status, totalAmount, items, createdAt, updatedAt]
    CreateOrderBody:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: ["string", "null"]
        type:
          type: string
          enum: [dine_in, takeout]
        tableId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemBody"
      required: [customerName, type, items]
    CreateOrderItemBody:
      type: object
      properties:
        menuItemId:
          type: integer
        quantity:
          type: integer
        notes:
          type: ["string", "null"]
      required: [menuItemId, quantity]
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, completed, cancelled]
      required: [status]
    Reservation:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        partySize:
          type: integer
        reservationDate:
          type: string
        reservationTime:
          type: string
        tableId:
          type: ["integer", "null"]
        tableNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, seated, completed, cancelled, no_show]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, customerName, customerPhone, partySize, reservationDate, reservationTime, status, createdAt]
    CreateReservationBody:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        partySize:
          type: integer
        reservationDate:
          type: string
        reservationTime:
          type: string
        tableId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required: [customerName, customerPhone, partySize, reservationDate, reservationTime]
    UpdateReservationStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, seated, completed, cancelled, no_show]
      required: [status]
...
    StaffMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        shift:
          type: ["string", "null"]
        isActive:
          type: boolean
        hireDate:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, role, isActive, createdAt]
    CreateStaffMemberBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        shift:
          type: ["string", "null"]
        isActive:
          type: boolean
        hireDate:
          type: ["string", "null"]
      required: [name, role, isActive]
...