# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hangout Alley Siargao Restaurant Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderBodyType } from "./createOrderBodyType";
import type { CreateOrderItemBody } from "./createOrderItemBody";

export interface CreateOrderBody {
  customerName: string;
  /** @nullable */
  customerPhone?: string | null;
  type: CreateOrderBodyType;
  /** @nullable */
  tableId?: number | null;
  /** @nullable */
  notes?: string | null;
  items: CreateOrderItemBody[];
}