# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hangout Alley Siargao Restaurant Management API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";
import type { OrderType } from "./orderType";

export interface Order {
  id: number;
  customerName: string;
  /** @nullable */
  customerPhone?: string | null;
  type: OrderType;
  status: OrderStatus;
  /** @nullable */
  tableId?: number | null;
  /** @nullable */
  tableNumber?: string | null;
  totalAmount: number;
  /** @nullable */
  notes?: string | null;
  items: OrderItem[];
  createdAt: string;
  updatedAt: string;
}